app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
python-frontmatter==1.1.0
markdown-it-py==3.0.0
argon2-cffi==25.1.0
itsdangerous==2.2.0
Pillow==11.3.0
pytest==8.4.1
httpx==0.28.1
email-validator==2.3.0
